third_party/WebKit/LayoutTests/external/wpt/webrtc/RTCPeerConnection-setRemoteDescription-offer-expected.txt

This is a testharness.js-based test.
FAIL setRemoteDescription with valid offer should succeed assert_not_equals: Expect session description to be defined got disallowed value undefined
FAIL setRemoteDescription multiple times should succeed assert_not_equals: Expect session description to be defined got disallowed value undefined
FAIL setRemoteDescription multiple times with different offer should succeed assert_not_equals: Expect session description to be defined got disallowed value undefined
FAIL setRemoteDescription(offer) with invalid SDP should reject with RTCError assert_equals: Expect error detail field to set to sdp-syntax-error expected (string) "sdp-syntax-error" but got (undefined) undefined
PASS setRemoteDescription(offer) from have-local-offer state should reject with InvalidStateError
Harness: the test ran to completion.
